+++ connect/ai/typesafe/client/src/main/kotlin/org/http4k/connect/typesafe/JevModels.kt
@@ -1,0 +1,9 @@
+package org.http4k.connect.typesafe
+
+import org.http4k.ai.model.ModelName
+
+object JevModels {
+    val JevLatest = ModelName.of("jev-latest")
+    val JevPreview = ModelName.of("jev-preview")
+    val Jev_1_13_0 = ModelName.of("jev-1.13.0")
+}