+++ connect/ai/typesafe/client/src/main/kotlin/org/http4k/connect/typesafe/TypeSafe.kt
@@ -1,0 +1,12 @@
+package org.http4k.connect.typesafe
+
+import dev.forkhandles.result4k.Result
+import org.http4k.connect.Http4kConnectApiClient
+import org.http4k.connect.RemoteFailure
+
+@Http4kConnectApiClient
+interface TypeSafe {
+    operator fun <R> invoke(action: TypeSafeAction<R>): Result<R, RemoteFailure>
+
+    companion object
+}