+++ connect/ai/typesafe/fake/src/test/kotlin/org/http4k/connect/typesafe/FakeTypeSafeChaosTest.kt
@@ -1,0 +1,9 @@
+package org.http4k.connect.typesafe
+
+import org.http4k.connect.FakeSystemContract
+import org.http4k.core.Method.GET
+import org.http4k.core.Request
+
+class FakeTypeSafeChaosTest : FakeSystemContract(FakeTypeSafe()) {
+    override val anyValid = Request(GET, "/v1/models")
+}