+++ connect/ai/typesafe/fake/src/test/kotlin/org/http4k/connect/typesafe/FakeTypeSafeTest.kt
@@ -1,0 +1,5 @@
+package org.http4k.connect.typesafe
+
+class FakeTypeSafeTest : TypeSafeContract {
+    override val typeSafe = FakeTypeSafe().client()
+}